pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["Ashraf  Ezz <Ashraf.Teleb99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"
django-rest-knox = "^4.2.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
django-filter = "^22.1"
celery = "^5.2.7"
redis = "^4.3.4"
django-celery-beat = "^2.4.0"
django-celery-results = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
